Production/computer_status_check_agent/pyproject.toml
[project]
name = "computer-status-agent"
version = "0.2.0"
description = "LangChain AI agent for automating Computer Status Check Linear tasks"
requires-python = ">=3.13"

dependencies = [
    # LangChain core + Anthropic (default provider)
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-core>=0.3",
    "anthropic>=0.50",

    # Pydantic for typed tool schemas
    "pydantic>=2.0",

    # HTTP clients
    "httpx>=0.27",

    # Date parsing for issue fields
    "python-dateutil>=2.9",

    # Slack SDK
    "slack-sdk>=3.27",

    # Environment / config
    "python-dotenv>=1.0",

    # Webhook listener (optional — needed for automated trigger)
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
openai = ["langchain-openai>=0.3"]
ollama = ["langchain-ollama>=0.3"]
all-providers = [
    "langchain-openai>=0.3",
    "langchain-ollama>=0.3",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
# Run agent against a single issue: uv run csa-run IT-1234
csa-run = "computer_status_agent.agent:main"

[tool.uv]
python = "3.13.3"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
strict = true
